requirements.txt
# Core Framework
streamlit>=1.28.0
langchain>=0.1.0
langgraph>=0.0.40
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
python-dateutil>=2.8.0

# Knowledge Base / RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0
requests>=2.31.0
langchain-tavily

